requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.3
elasticsearch==6.3.1
Flask==1.0.2
Flask-WTF==0.14.2
hexdump==3.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
libesedb-python==20181229
MarkupSafe==1.1.1
more-itertools==6.0.0
pkg-resources==0.0.0
pluggy==0.9.0
py==1.8.0
pytest==4.3.1
pytest-cov==2.6.1
python-dateutil==2.8.0
python-evtx==0.6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.1
WTForms==2.2.1
xmltodict==0.12.0
